+++ CHANGES.md
@@ -5,9 +5,10 @@ version ?????????
 ----------------------
 - Bugs fix.
 - Reduce the memory footprint of classes in the model package.
-- Introduce a static partitioning algorithm that split a problem from disjoint set of nodes
+- StaticPartitioning: an experimental partitioning algorithm that splits
+  a problem from disjoint set of nodes
 - API changes in Model, Instance, Mapping
-- new constraint: MaxOnlines
+- new constraint: MaxOnline
 
 version 0.31 - 16 Jul 2013
 ----------------------